Move action messages to actions
